put change password and logout in same data source
