+++ lib/features/user_profile/data/data_source/user_profile_data_source_repo_imp.dart
@@ -50,3 +50,56 @@ class UserPofileDataSourceRepoImp implements UserProfileDataSourceRepo {
     }
   }
 }
+
+@Injectable(as: ChangePasswordDataSource)
+class ChangePasswordDataSourceImp implements ChangePasswordDataSource {
+  AuthService authService;
+
+  ChangePasswordDataSourceImp(this.authService);
+
+  @override
+  Future<Either<ServerFailure, void>> changePassword({required String oldPassword, required String newPassword, required String reNewPassword}) async{
+    try{
+      Response response =await authService.changePassword(oldPassword: oldPassword,
+          newPassword: newPassword,
+          reNewPassword: reNewPassword);
+      if(response.statusCode! >= 200 && response.statusCode! < 300){
+
+        return Right(null);
+      }
+      else{
+        return left(ServerFailure(errorMessage: response.data['message']));
+      }
+    }on DioException catch(e){
+      print("=======================Error ======");
+      print(e.toString());
+      return left(ServerFailure.fromDioException(e));
+    }
+  }
+
+
+
+}
+
+
+@Injectable(as: LogoutDataSource)
+class LogoutDataSourceImp implements LogoutDataSource {
+  AuthService authService;
+
+  LogoutDataSourceImp(this.authService);
+
+  @override
+  Future<Either<ServerFailure, void>> logout() async {
+    try {
+      Response response = await authService.logout();
+      if (response.statusCode! >= 200 && response.statusCode! < 300) {
+        return Right(null);
+      } else {
+        return left(ServerFailure(errorMessage: response.data['message']));
+      }
+    } on DioException catch (e) {
+      return left(ServerFailure.fromDioException(e));
+    }
+  }
+}
+
